pyproject.toml
[project]
name = "gws-wrapper"
version = "0.1.0"
description = "Google Workspace CLI Wrapper"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.2",
    "dynaconf>=3.2.13",
    "loguru>=0.7.3",
    "pydantic>=2.13.2",
    "python-dateutil>=2.9.0.post0",
]

[project.scripts]
gws-wrap = "gws_wrapper.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gws_wrapper"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
]
